+++ src/main/java/org/plumelib/util/CollectionsPlume.java
@@ -174,6 +174,8 @@ public final class CollectionsPlume {
    *
    * <pre>import static org.plumelib.util.CollectionsPlume.mapList;</pre>
    *
+   * This method is just like {@link #transform}, but with the arguments in the other order.
+   *
    * @param <FROM> the type of elements of the given collection
    * @param <TO> the type of elements of the result list
    * @param f a function
@@ -181,6 +183,25 @@ public final class CollectionsPlume {
    * @return a list of the results of applying {@code f} to the elements of {@code list}
    */
   public static <FROM, TO> List<TO> mapList(Function<? super FROM, TO> f, Collection<FROM> c) {
+    return c.stream().map(f).collect(Collectors.toList());
+  }
+
+  /**
+   * Applies the function to each element of the given collection, producing a list of the results.
+   *
+   * <p>The point of this method is to make mapping operations more concise. Import it with
+   *
+   * <pre>import static org.plumelib.util.CollectionsPlume.transform;</pre>
+   *
+   * This method is just like {@link #mapList}, but with the arguments in the other order.
+   *
+   * @param <FROM> the type of elements of the given collection
+   * @param <TO> the type of elements of the result list
+   * @param c a collection
+   * @param f a function
+   * @return a list of the results of applying {@code f} to the elements of {@code list}
+   */
+  public static <FROM, TO> List<TO> transform(Collection<FROM> c, Function<? super FROM, TO> f) {
     return c.stream().map(f).collect(Collectors.toList());
   }